+++ CHANGELOG.md
@@ -10,7 +10,7 @@ All notable changes to this project will be documented in this file.
 - Fixed bug [#25](https://github.com/annaet/cypress-tags/issues/25) where tagged describe blocks were being ignored completely when the `CYPRESS_INCLUDE_TAGS` env var included tags that did not match the describe block. Now `cypress-tags` should fall through to the inner test blocks and check those for tags instead.
 - Update README to explain why using Cypress's `--env` CLI argument does not work as referenced in [#35](https://github.com/annaet/cypress-tags/issues/35).
 - Fixed bug [#47](https://github.com/annaet/cypress-tags/issues/47) where `xit` and `xdescribe` being used were causing tests to fail.
-
+- Adds new environment variables to allow a user to boolean AND tags together when including or excluding as in [#34](https://github.com/annaet/cypress-tags/issues/34).
 
 ## [0.0.21] - 2021-03-28